extract @commons-systems/criticalcssutil from htmlutil
Move critical-css.ts, critters.d.ts, and the test suite out of htmlutil
into a new standalone package @commons-systems/criticalcssutil. Update
fellspiral and landing to depend on criticalcssutil and import
inlineCriticalCss from the new package. Remove the ./critical-css export
from htmlutil so it only exports escapeHtml going forward.
