+++ criticalcssutil/eslint.config.js
@@ -1,0 +1,1 @@
+export { default } from "@commons-systems/config/eslint";
