+++ criticalcssutil/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@commons-systems/htmlutil",
+  "name": "@commons-systems/criticalcssutil",
   "private": true,
   "version": "0.0.0",
   "type": "module",
