+++ criticalcssutil/test/critical-css.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { mkdtemp, mkdir, writeFile, readFile, rm } from "node:fs/promises";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
-import { inlineCriticalCss } from "../src/critical-css";
+import { inlineCriticalCss } from "../src/index";
 
 const MINIMAL_CSS = `
 body { margin: 0; font-family: sans-serif; }
