+++ criticalcssutil/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@commons-systems/config/tsconfig.lib.json",
+  "include": ["src", "test"]
+}
