+++ criticalcssutil/vite.config.ts
@@ -1,0 +1,3 @@
+import { createLibConfig } from "@commons-systems/config/vite";
+
+export default createLibConfig();
