+++ fellspiral/package.json
@@ -18,6 +18,7 @@
     "@commons-systems/blog": "*",
     "@commons-systems/router": "*",
     "@commons-systems/firebaseutil": "*",
+    "@commons-systems/criticalcssutil": "*",
     "@commons-systems/htmlutil": "*",
     "@commons-systems/firestoreutil": "*",
     "@commons-systems/style": "*",
