+++ landing/scripts/critical-css.ts
@@ -1,5 +1,5 @@
 import { dirname, join } from "node:path";
-import { inlineCriticalCss } from "@commons-systems/htmlutil/critical-css";
+import { inlineCriticalCss } from "@commons-systems/criticalcssutil";
 
 const distDir = join(dirname(new URL(import.meta.url).pathname), "..", "dist");
 await inlineCriticalCss(distDir);
