+++ package.json
@@ -8,6 +8,7 @@
     "authutil",
     "blog",
     "budget",
+    "criticalcssutil",
     "errorutil",
     "fellspiral",
     "firebaseutil",
